% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_custom.R
\name{mean_custom}
\alias{mean_custom}
\title{This function calculates the mean of a numeric vector with custom options.}
\usage{
mean_custom(x, na.rm = FALSE)
}
\arguments{
\item{x}{any character vector}

\item{na.rm}{Logical; should missing values be removed?}
}
\value{
a numeric number
}
\description{
This function calculates the mean of a numeric vector with custom options.
}
\examples{
x <- c(1,2,3,4,5)
mean_custom(x)
}
